requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.1
httpx>=0.27.0
google-genai>=1.20.0
elevenlabs>=1.6.0
